+++ python/mqt/qmap/_compat/typing.py
@@ -9,12 +9,7 @@
 from __future__ import annotations
 
 import sys
-from typing import TYPE_CHECKING
-
-if sys.version_info >= (3, 10):
-    from typing import TypeAlias
-else:
-    from typing_extensions import TypeAlias
+from typing import TYPE_CHECKING, TypeAlias
 
 if sys.version_info >= (3, 11):
     from typing import Self