+++ python/mqt/qmap/plugins/qiskit/sc/subarchitectures.py
@@ -17,7 +17,7 @@ import contextlib
 import pickle
 from itertools import combinations
 from pathlib import Path
-from typing import TYPE_CHECKING, Optional
+from typing import TYPE_CHECKING
 
 import rustworkx as rx
 import rustworkx.visualization as rxviz
@@ -35,7 +35,7 @@ if TYPE_CHECKING:
 
 
 with contextlib.suppress(TypeError):
-    Graph: TypeAlias = rx.PyGraph[int, Optional[int]]
+    Graph: TypeAlias = rx.PyGraph[int, int | None]
 
 PartialOrder: TypeAlias = dict[tuple[int, int], set[tuple[int, int]]]